requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.13.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.27.0
google-api-python-client>=2.150.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1
pytest>=8.3.0
pytest-asyncio>=0.24.0

